refactor: 커스텀 exception handler 중복코드 정리
