+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/ZzimkkongException.java
@@ -1,0 +1,16 @@
+package com.woowacourse.zzimkkong.exception;
+
+import org.springframework.http.HttpStatus;
+
+public class ZzimkkongException extends RuntimeException {
+    private final HttpStatus status;
+
+    public ZzimkkongException(final String message, final HttpStatus status) {
+        super(message);
+        this.status = status;
+    }
+
+    public HttpStatus getStatus() {
+        return status;
+    }
+}
