+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/map/MapException.java
@@ -1,16 +1,10 @@
 package com.woowacourse.zzimkkong.exception.map;
 
+import com.woowacourse.zzimkkong.exception.ZzimkkongException;
 import org.springframework.http.HttpStatus;
 
-public class MapException extends RuntimeException {
-    private final HttpStatus status;
-
+public class MapException extends ZzimkkongException {
     public MapException(final String message, final HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-
-    public HttpStatus getStatus() {
-        return status;
+        super(message, status);
     }
 }
