+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/member/MemberException.java
@@ -1,16 +1,10 @@
 package com.woowacourse.zzimkkong.exception.member;
 
+import com.woowacourse.zzimkkong.exception.ZzimkkongException;
 import org.springframework.http.HttpStatus;
 
-public class MemberException extends RuntimeException {
-    private final HttpStatus status;
-
+public class MemberException extends ZzimkkongException {
     public MemberException(final String message, final HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-
-    public HttpStatus getStatus() {
-        return status;
+        super(message, status);
     }
 }
