+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/reservation/ReservationException.java
@@ -1,16 +1,10 @@
 package com.woowacourse.zzimkkong.exception.reservation;
 
+import com.woowacourse.zzimkkong.exception.ZzimkkongException;
 import org.springframework.http.HttpStatus;
 
-public class ReservationException extends RuntimeException {
-    private final HttpStatus status;
-
+public class ReservationException extends ZzimkkongException {
     public ReservationException(final String message, final HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-
-    public HttpStatus getStatus() {
-        return status;
+        super(message, status);
     }
 }
