+++ backend/src/main/java/com/woowacourse/zzimkkong/exception/space/SpaceException.java
@@ -1,16 +1,10 @@
 package com.woowacourse.zzimkkong.exception.space;
 
+import com.woowacourse.zzimkkong.exception.ZzimkkongException;
 import org.springframework.http.HttpStatus;
 
-public class SpaceException extends RuntimeException {
-    private final HttpStatus status;
-
+public class SpaceException extends ZzimkkongException {
     public SpaceException(final String message, final HttpStatus status) {
-        super(message);
-        this.status = status;
-    }
-
-    public HttpStatus getStatus() {
-        return status;
+        super(message, status);
     }
 }
